[v1.3.36] update imports
