--- src/pages/layouts/BaseLayoutStyles.ts
@@ -1,20 +1,0 @@
-export const headerStyle = (colorBgContainer: string) => ({
-  paddingLeft: 7,
-  background: colorBgContainer,
-  margin: 0,
-  height: 30,
-  display: "flex"
-});
-export const buttonSiderStyle = {
-  fontSize: "14px",
-  width: 30,
-  height: 30,
-  marginTop: 1
-};
-export const contentStyle = (
-  colorBgContainer: string,
-  borderRadiusLG: number
-) => ({
-  background: colorBgContainer,
-  borderRadius: borderRadiusLG,
-});
