+++ src/pages/layouts/MainContainer.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import PageTitle from "../../pages/components/PageTitle";
+import PageTitle from "../components/PageTitle";
 import { Button, Input, Space } from "antd";
 import { IoIosSearch } from "react-icons/io";
 import Strings from "../../utils/localizations/Strings";
-import Loading from "../components/Loading";
-import BackButton from "../components/BackButton";
+import Loading from "../../pagesRedesign/components/Loading";
+import BackButton from "../../pagesRedesign/components/BackButton";
 
 interface MainContainerProps {
   title: string;
