+++ src/pages/priority/PrioritiesPage.tsx
@@ -6,7 +6,7 @@ import Strings from "../../utils/localizations/Strings";
 import { useLocation, useNavigate } from "react-router-dom";
 import PaginatedList from "../components/PaginatedList";
 import { UnauthorizedRoute } from "../../utils/Routes";
-import MainContainer from "../../pagesRedesign/layout/MainContainer";
+import MainContainer from "../layouts/MainContainer";
 import useCurrentUser from "../../utils/hooks/useCurrentUser";
 import PriorityForm, { PriorityFormType } from "./components/PriorityForm";
 import PriorityCard from "./components/PriorityCard";
