+++ src/pages/tags/TagsPage.tsx
@@ -4,7 +4,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { useGetCardsMutation } from "../../services/cardService";
 import { CardInterface } from "../../data/card/card";
 import { UnauthorizedRoute } from "../../utils/Routes";
-import MainContainer from "../../pagesRedesign/layout/MainContainer";
+import MainContainer from "../layouts/MainContainer";
 import useCurrentUser from "../../utils/hooks/useCurrentUser";
 import { useDebounce } from "use-debounce";
 import { handleErrorNotification } from "../../utils/Notifications";
