+++ src/pages/user/UsersPage.tsx
@@ -8,7 +8,7 @@ import { useGetUsersWithPositionsMutation } from "../../services/userService";
 import { UserCardInfo } from "../../data/user/user";
 import { useLocation, useNavigate } from "react-router-dom";
 import { UnauthorizedRoute } from "../../utils/Routes";
-import MainContainer from "../../pagesRedesign/layout/MainContainer";
+import MainContainer from "../layouts/MainContainer";
 import useCurrentUser from "../../utils/hooks/useCurrentUser";
 import PaginatedList from "../components/PaginatedList";
 import AnatomyNotification from "../components/AnatomyNotification";
